Add strict parsing mode
The default parser doesn't consider begin, end, version, or fn
fields while parsing. This leads to odd behaviour if you give
it an invalid vCard - such as one without an begin or end line;
without a version or full name - where it parses it as if it
were valid. This problem is most obvious if you give the parser
a file that contains multiple cards, it will merge all the
cards into a single large card.